[expo-image][ios] Add class for types & convert functions
