+++ packages/expo-image/ios/EXImageView.h
@@ -1,18 +1,13 @@
 // Copyright 2020-present 650 Industries. All rights reserved.
 
 #import <SDWebImage/SDWebImage.h>
+#import <expo-image/EXImageView.h>
+#import <expo-image/EXImageTypes.h>
 #import <React/RCTComponent.h>
 #import <React/UIView+React.h>
 #import <React/RCTResizeMode.h>
 
 @class RCTBridge;
-
-typedef NS_ENUM(NSInteger, EXImageCacheTypeEnum) {
-  EXImageCacheUnknown = 0,
-  EXImageCacheNone,
-  EXImageCacheDisk,
-  EXImageCacheMemory
-};
 
 @interface EXImageView : SDAnimatedImageView
